Fix graph crash with partial conditions
Graph.rep() assumes that every callable has a `__name__` attribute.
However, callables like `functools.partial` and other classes with a
`__call__()` method have no `__name__`.  This patch modifies
`Graph.rep()` to first check whether the condition is a string and
return it, then try to return `f.__name__` (that is the current
behavior) and if that fails, then it calls `str()` on the argument.

`str()` should never fail and it is now the responsibility of the user
to return something meaningful from `__str__()`.